Move type defenitions
