+++ apps/judicial-system/web/src/routes/Prosecutor/SharedComponents/DefendantInfo/DefendantInfo.tsx
@@ -146,7 +146,7 @@ const DefendantInfo: React.FC<Props> = (props) => {
             })
           }}
           onBlur={async (evt) => {
-            if (person && person.items.length === 1) {
+            if (person && person.items.length === 1 && !error) {
               onChange(defendant.id, {
                 nationalId: evt.target.value,
                 name: person.items[0].name,
